MessagesRepository implemented, UsersRepository fully covered, ChatsRepository coverage in process

+++ DotNetMessenger.DataLayer.SqlServer/UsersRepository.cs
@@ -1,4 +1,5 @@
-﻿using System.Data;
+﻿using System;
+using System.Data;
 using System.Data.SqlClient;
 using DotNetMessenger.Model;
 using DotNetMessenger.Model.Enums;
@@ -8,20 +9,31 @@ namespace DotNetMessenger.DataLayer.SqlServer
     public class UsersRepository : IUsersRepository
     {
         private readonly string _connectionString;
-        private readonly IChatsRepository _chatsRepository;
+        public IChatsRepository ChatsRepository { get; set; }
+        
 
         public UsersRepository(string connectionString, IChatsRepository chatsRepository)
         {
             _connectionString = connectionString;
-            _chatsRepository = chatsRepository;
+            ChatsRepository = chatsRepository;
+        }
+
+        public UsersRepository(string connectionString)
+        {
+            _connectionString = connectionString;
         }
 
         public User CreateUser(string userName, string hash)
         {
+            if (userName == string.Empty || hash == string.Empty)
+                return null;
+            
             using (var connection = new SqlConnection(_connectionString))
             {
                 connection.Open();
 
+                if (SqlHelper.DoesFieldValueExist(connection, "Users", "Username", userName, SqlDbType.VarChar, userName.Length))
+                    return null;
                 using (var transaction = connection.BeginTransaction())
                 {
                     User user;
@@ -58,29 +70,21 @@ namespace DotNetMessenger.DataLayer.SqlServer
 
         public void DeleteUser(int userId)
         {
+            // Default (deleted) user check
+            if (userId == 0)
+                return;
+            
             using (var connection = new SqlConnection(_connectionString))
             {
                 connection.Open();
 
+                //Check if id exists
+                if (!SqlHelper.DoesFieldValueExist(connection, "Users", "ID", userId, SqlDbType.Int))
+                    return;
                 // Just delete the user, everything else will cascade delete
                 using (var command = connection.CreateCommand())
                 {
                     command.CommandText = "DELETE FROM [Users] WHERE [ID] = @userId";
-                    command.Parameters.AddWithValue("@userId", userId);
-                    command.ExecuteNonQuery();
-                }
-            }
-        }
-
-        public void DeleteUserInfo(int userId)
-        {
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                connection.Open();
-
-                using (var command = connection.CreateCommand())
-                {
-                    command.CommandText = "DELETE FROM [UserInfos] WHERE [UserID] = @userId";
                     command.Parameters.AddWithValue("@userId", userId);
                     command.ExecuteNonQuery();
                 }
@@ -93,6 +97,8 @@ namespace DotNetMessenger.DataLayer.SqlServer
             {
                 connection.Open();
 
+                if (!SqlHelper.DoesFieldValueExist(connection, "Users", "ID", userId, SqlDbType.Int))
+                    return null;
                 using (var command = connection.CreateCommand())
                 {
                     command.CommandText = "SELECT [ID], [Username], [Password] FROM [Users] WHERE [ID] = @userId";
@@ -102,16 +108,33 @@ namespace DotNetMessenger.DataLayer.SqlServer
                     using (var reader = command.ExecuteReader())
                     {
                         reader.Read();
-                        user.Id = reader.GetInt32(reader.GetOrdinal("[ID]"));
-                        user.Hash = reader.GetString(reader.GetOrdinal("[Password]"));
-                        user.Username = reader.GetString(reader.GetOrdinal("[Username]"));
+                        user.Id = reader.GetInt32(reader.GetOrdinal("ID"));
+                        user.Hash = reader.GetString(reader.GetOrdinal("Password"));
+                        user.Username = reader.GetString(reader.GetOrdinal("Username"));
                     }
-                    user.Chats = _chatsRepository.GetUserChats(user.Id);
+                    user.Chats = ChatsRepository.GetUserChats(user.Id);
                     user.UserInfo = GetUserInfo(user.Id);
                     return user;
                 }
             }
         }
+
+        public void DeleteUserInfo(int userId)
+        {
+            using (var connection = new SqlConnection(_connectionString))
+            {
+                connection.Open();
+                if (!SqlHelper.DoesFieldValueExist(connection, "Users", "ID", userId, SqlDbType.Int)
+                    || !SqlHelper.DoesFieldValueExist(connection, "UserInfos", "UserID", userId, SqlDbType.Int))
+                    return;
+                using (var command = connection.CreateCommand())
+                {
+                    command.CommandText = "DELETE FROM [UserInfos] WHERE [UserID] = @userId";
+                    command.Parameters.AddWithValue("@userId", userId);
+                    command.ExecuteNonQuery();
+                }
+            }
+        } 
 
         public UserInfo GetUserInfo(int userId)
         {
@@ -119,6 +142,8 @@ namespace DotNetMessenger.DataLayer.SqlServer
             {
                 connection.Open();
 
+                if (!SqlHelper.DoesFieldValueExist(connection, "UserInfos", "UserID", userId, SqlDbType.Int))
+                    return null;
                 using (var command = connection.CreateCommand())
                 {
                     command.CommandText = "SELECT [LastName], [FirstName], [Phone], [Email], [DateOfBirth], [Avatar]" +
@@ -131,12 +156,12 @@ namespace DotNetMessenger.DataLayer.SqlServer
                         reader.Read();
                         return new UserInfo
                         {
-                            LastName = reader.GetString(reader.GetOrdinal("[LastName]")),
-                            FirstName = reader.GetString(reader.GetOrdinal("[FirstName]")),
-                            Phone = reader.GetString(reader.GetOrdinal("[Phone]")),
-                            Email = reader.GetString(reader.GetOrdinal("[Email]")),
-                            DateOfBirth = reader.GetDateTime(reader.GetOrdinal("[DateOfBirth]")),
-                            Avatar = reader[reader.GetOrdinal("[Avatar]")] as byte[]
+                            LastName = reader.IsDBNull(reader.GetOrdinal("LastName")) ? null : reader.GetString(reader.GetOrdinal("LastName")),
+                            FirstName = reader.IsDBNull(reader.GetOrdinal("FirstName")) ? null : reader.GetString(reader.GetOrdinal("FirstName")),
+                            Phone = reader.IsDBNull(reader.GetOrdinal("Phone")) ? null : reader.GetString(reader.GetOrdinal("Phone")),
+                            Email = reader.IsDBNull(reader.GetOrdinal("Email")) ? null : reader.GetString(reader.GetOrdinal("Email")),
+                            DateOfBirth = reader.IsDBNull(reader.GetOrdinal("DateOfBirth")) ? DateTime.MinValue : reader.GetDateTime(reader.GetOrdinal("DateOfBirth")),
+                            Avatar = reader.IsDBNull(reader.GetOrdinal("Avatar")) ? null : reader[reader.GetOrdinal("Avatar")] as byte[]
                         };
                     }
                 }
@@ -145,28 +170,83 @@ namespace DotNetMessenger.DataLayer.SqlServer
 
         public void SetUserInfo(int userId, UserInfo userInfo)
         {
+            if (userId == 0)
+                return;
+            if (userInfo == null)
+                return;
+
             using (var connection = new SqlConnection(_connectionString))
             {
                 connection.Open();
 
+                if (!SqlHelper.DoesFieldValueExist(connection, "Users", "ID", userId, SqlDbType.Int))
+                    return;
                 using (var command = connection.CreateCommand())
                 {
-                    /* TODO: CHECK IF NO USERINFO EXISTS, INSERT IF YES */
-                    command.CommandText = "UPDATE [UserInfos] SET [LastName] = @lastName, [FirstName] = @firstName, " +
+                    if (SqlHelper.DoesFieldValueExist(connection, "UserInfos", "UserID", userId, SqlDbType.Int))
+                    {
+                        command.CommandText = "UPDATE [UserInfos] SET [LastName] = @lastName, [FirstName] = @firstName, " +
                                           "[Phone] = @phone, [Email] = @email, [DateOfBirth] = @dateOfBirth, " +
                                           "[Avatar] = @avatar WHERE [UserID] = @userId";
+                    }
+                    else
+                    {
+                        command.CommandText =
+                            "INSERT INTO [UserInfos] ([UserID], [LastName], [FirstName], [Phone], [Email], [DateOfBirth], " +
+                            "[Avatar]) VALUES (@userId, @lastName, @firstName, @phone, @email, @dateOfBirth, @avatar)";
+                    }
 
                     command.Parameters.AddWithValue("@userId", userId);
-                    command.Parameters.AddWithValue("@lastName", userInfo.LastName);
-                    command.Parameters.AddWithValue("@firstName", userInfo.FirstName);
-                    command.Parameters.AddWithValue("@phone", userInfo.Phone);
-                    command.Parameters.AddWithValue("@email", userInfo.Email);
+
+                    if (userInfo.LastName == null)
+                    {
+                        command.Parameters.AddWithValue("@lastName", DBNull.Value);
+                    }
+                    else
+                    {
+                        command.Parameters.AddWithValue("@lastName", userInfo.LastName);
+                    }
+
+                    if (userInfo.FirstName == null)
+                    {
+                        command.Parameters.AddWithValue("@firstName", DBNull.Value);
+                    }
+                    else
+                    {
+                        command.Parameters.AddWithValue("@firstName", userInfo.FirstName);
+                    }
+
+                    if (userInfo.Phone == null)
+                    {
+                        command.Parameters.AddWithValue("@phone", DBNull.Value);
+                    }
+                    else
+                    {
+                        command.Parameters.AddWithValue("@phone", userInfo.Phone);
+                    }
+
+                    if (userInfo.Email == null)
+                    {
+                        command.Parameters.AddWithValue("@email", DBNull.Value);
+                    }
+                    else
+                    {
+                        command.Parameters.AddWithValue("@email", userInfo.Email);
+                    }
+
                     command.Parameters.AddWithValue("@dateOfBirth", userInfo.DateOfBirth);
 
-                    var avatar =
-                        new SqlParameter("@avatar", SqlDbType.VarBinary, userInfo.Avatar.Length)
-                            { Value = userInfo.Avatar };
-                    command.Parameters.Add(avatar);
+                    if (userInfo.Avatar == null)
+                    {
+                        command.Parameters.Add(new SqlParameter("@avatar", SqlDbType.VarBinary) {Value = DBNull.Value});
+                    }
+                    else
+                    {
+                        var avatar =
+                            new SqlParameter("@avatar", SqlDbType.VarBinary, userInfo.Avatar.Length)
+                                {Value = userInfo.Avatar};
+                        command.Parameters.Add(avatar);
+                    }
 
                     command.ExecuteNonQuery();
                 }
@@ -175,22 +255,37 @@ namespace DotNetMessenger.DataLayer.SqlServer
 
         public User PersistUser(User user)
         {
+            if (user == null)
+                return null;
+            // Check for default user change
+            if (user.Id == 0)
+                return null;
+            
             using (var connection = new SqlConnection(_connectionString))
             {
                 connection.Open();
 
+                var idExists = SqlHelper.DoesFieldValueExist(connection, "Users", "ID", user.Id, SqlDbType.Int);
+                // If ID does not exist, but username exists, then we cannot create a new user!
+                if (!idExists && SqlHelper.DoesFieldValueExist(connection, "Users", "Username", user.Username, SqlDbType.VarChar,
+                        user.Username.Length))
+                    return null;
                 using (var command = connection.CreateCommand())
                 {
-                    
-                    command.CommandText =
-                        "UPDATE [Users] SET [Username] = @userName, [Password] = @password WHERE [ID] = @userId";
+                    command.CommandText = idExists ? 
+                        "UPDATE [Users] SET [Username] = @userName, [Password] = @password WHERE [ID] = @userId" 
+                        : "INSERT INTO [Users] ([Username], [Password]) OUTPUT INSERTED.[ID] VALUES (@userName, @password)";
 
-                    command.Parameters.AddWithValue("@userId", user.Id);
+                    if (idExists)
+                        command.Parameters.AddWithValue("@userId", user.Id);
                     command.Parameters.AddWithValue("@userName", user.Username);
                     command.Parameters.AddWithValue("@password", user.Hash);
 
-                    command.ExecuteNonQuery();
-                    
+                    if (idExists)
+                        command.ExecuteNonQuery();
+                    else
+                        user.Id = (int) command.ExecuteScalar();    
+
                     if (user.UserInfo != null)
                         SetUserInfo(user.Id, user.UserInfo);
                     return user;
@@ -199,113 +294,17 @@ namespace DotNetMessenger.DataLayer.SqlServer
             }
         }
 
-        public ChatUserInfo GetChatSpecificInfo(int userId, int chatId)
-        {
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                connection.Open();
-
-                var chatUserInfo = new ChatUserInfo();
-                int roleId;
-
-                using (var command = connection.CreateCommand())
-                {
-                    command.CommandText =
-                        "SELECT [Nickname], [UserRole] FROM [ChatUserInfos] WHERE [UserID] = @userId AND [ChatID] = @chatId";
-
-                    command.Parameters.AddWithValue("@userId", userId);
-                    command.Parameters.AddWithValue("@chatId", chatId);
-
-                    using (var reader = command.ExecuteReader())
-                    {
-                        reader.Read();
-                        chatUserInfo.Nickname = reader.GetString(reader.GetOrdinal("[Nickname]"));
-                        roleId = reader.GetInt32(reader.GetOrdinal("[UserRole]"));
-                    }
-                }
-
-                using (var command = connection.CreateCommand())
-                {
-                    command.CommandText =
-                        "SELECT [Name], [ReadPerm], [WritePerm], [ChatInfoPerm], [AttachPerm], [ManageUsersPerm]" +
-                        " FROM [UserRoles] WHERE [ID] = @roleId";
-                    command.Parameters.AddWithValue("@roleId", roleId);
-
-                    using (var reader = command.ExecuteReader())
-                    {
-                        reader.Read();
-                        chatUserInfo.Role = new UserRole
-                        {
-                            RoleType = (UserRoles) roleId,
-                            ReadPerm = reader.GetBoolean(reader.GetOrdinal("[ReadPerm]")),
-                            WritePerm = reader.GetBoolean(reader.GetOrdinal("[WritePerm]")),
-                            ChatInfoPerm = reader.GetBoolean(reader.GetOrdinal("[ChatInfoPerm]")),
-                            AttachPerm = reader.GetBoolean(reader.GetOrdinal("[AttachPerm]")),
-                            ManageUsersPerm = reader.GetBoolean(reader.GetOrdinal("[ManageUsersPerm]")),
-                            RoleName = reader.GetString(reader.GetOrdinal("[Name]"))
-                        };
-                    }
-                }
-
-                return chatUserInfo;
-            }
-        }
-
-        public void SetChatSpecificInfo(int userId, int chatId, ChatUserInfo userInfo, bool updateRole = false)
-        {
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                connection.Open();
-
-                using (var command = connection.CreateCommand())
-                {
-                    /* TODO: CHECK FOR MISSING CHAT USER INFO, IF MISSING, ADD */
-                    command.CommandText = "UPDATE [ChatUserInfos] SET [Nickname] = @nickName";
-                    if (updateRole)
-                        command.CommandText += ", [UserRole] = @userRole";
-                    command.CommandText += " WHERE [UserID] = @userId AND [ChatID] = @chatId";
-
-                    command.Parameters.AddWithValue("@nickName", userInfo.Nickname);
-                    command.Parameters.AddWithValue("@chatId", chatId);
-                    command.Parameters.AddWithValue("@userId", userId);
-                    if (updateRole)
-                    {
-                        command.Parameters.AddWithValue("@userRole", (int)(userInfo.Role?.RoleType ?? UserRoles.Regular));
-                    }
-
-                    command.ExecuteNonQuery();
-                }
-            }
-        }
-
-        public ChatUserInfo SetChatSpecificRole(int userId, int chatId, UserRoles userRole)
-        {
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                connection.Open();
-
-                using (var command = connection.CreateCommand())
-                {
-                    /* TODO: CHECK FOR MISSING CHAT USER INFO, IF MISSING, ADD */
-                    command.CommandText = "UPDATE [ChatUserInfos] SET [UserRole] = @userRole " +
-                                          "WHERE [UserID] = @userId AND [ChatID] = @chatId";
-
-                    command.Parameters.AddWithValue("@userRole", (int) userRole);
-                    command.Parameters.AddWithValue("@chatId", chatId);
-                    command.Parameters.AddWithValue("@userId", userId);
-
-                    command.ExecuteNonQuery();
-                    return GetChatSpecificInfo(userId, chatId);
-                }
-            }
-        }
-
         public void SetPassword(int userId, string newHash)
         {
+            if (userId == 0)
+                return;
+            
             using (var connection = new SqlConnection(_connectionString))
             {
                 connection.Open();
 
+                if (!SqlHelper.DoesFieldValueExist(connection, "Users", "ID", userId, SqlDbType.Int))
+                    return;
                 using (var command = connection.CreateCommand())
                 {
                     command.CommandText = "UPDATE [Users] SET [Password] = @password WHERE [ID] = @userId";
